Several changes have been made:
Added files:
- main.py, the file which runs all the other processes;
- data_parser.py, which will manage the data parsing;
- data_publisher.py, which will handle the communication with mqtt broker.

Serial_Port_Reader.py has been renamed to serial_port_reader.py and some changes have been made:
- threads are now created from a function with dynamic arguments;
- buffer has been splitted in 3 buffers to split the data gathered from the different data sources;
- added some error handling;
- added pipes for inter process communication.

The different processes now send data to each other through pipes (multiprocessing library).
The pipes are created in the main file, and the writing pipe must be closed before exiting the function or it will remain pending (no need to close reading pipes).
Some test should be done to understand if it's better to use processes or threads on the raspberry.
It is also possible to create an asyncio loop and manage all the processes as asyncronous functions (which can also spawn threads).